chore(devex): move digest summary dataclasses to facade contracts
The facade isolation repo check rejects facade/api.py re-exporting ExceptionSummary and CrashFreeSummary from backend/weekly_digest.py. Move both dataclasses into facade/contracts.py (pydantic frozen dataclasses, matching the file) and import them from there in weekly_digest, the facade, the temporal activities, and tests.

SOURCE_MAPS_DOCS_URL moves along with them: weekly_digest now imports from the facade package, so api.py reading the constant off weekly_digest at module level would break when weekly_digest is the first module imported.

Generated-By: PostHog Desktop
Task-Id: 3cb06508-6ac6-4768-a330-4495ac21001e
