+++ products/error_tracking/backend/temporal/weekly_digest/activities.py
@@ -17,6 +17,7 @@ from posthog.temporal.common.heartbeat_sync import HeartbeaterSync
 from posthog.user_permissions import UserPermissions
 
 from products.error_tracking.backend import weekly_digest
+from products.error_tracking.backend.facade.contracts import ExceptionSummary
 from products.error_tracking.backend.temporal.weekly_digest.types import (
     CleanupDigestOrgsInputs,
     GetDigestOrgsInputs,
@@ -119,7 +120,7 @@ def _send_org_digest(inputs: SendOrgDigestInputs, attempt: int) -> SendOrgDigest
     # counts: unfiltered counts can permanently enroll a user onto a project whose digest builds empty
     # (auto-select is a one-shot decision). Only computed when the org actually has a first-time user.
     setting_key = weekly_digest.DIGEST_PROJECT_SETTING_KEY
-    autoselect_counts: dict[int, weekly_digest.ExceptionSummary] = {}
+    autoselect_counts: dict[int, ExceptionSummary] = {}
     # Kept for Pass 2: the 14-day row query is the most expensive thing this activity does, so a team
     # ranked here shouldn't pay for it again when its digest is built.
     daily_rows_by_team: dict[int, list] = {}
